# Optimize transform propagation for hidden 3d objects