When visual instances are hidden, there is no need to update their global transforms and send these to the visual server. This only needs to be done when the objects are reshown.

## Explanation
I'm currently looking at optimizing performance of hidden objects, as I've noticed they surprisingly can slow down frame rates significantly. While #45576 deals with the case of static objects, there is still a performance problem when hidden 3d objects are moved.

### Benchmarks:
_0 objects_
6500 fps

_10,000 objects, hidden and static (with #45576)_
5500 fps

_10,000 objects, hidden but moved each frame via moving parent node_
158 fps

_10,000 objects, hidden but moved each frame via moving parent node, with this PR_
360 fps

Clearly the drop with dynamic objects is very significant, and should in theory be totally unnecessary.

Profiling revealed a lot of time was spent calling `get_global_transform()` and sending the transform to the visual server.

This example PR keeps a local flag for visibility state (something we should probably be doing anyway, as `is_visible_in_tree` can be expensive), doesn't update these transforms when the node is not visible.

This increases fps to 360 fps. While an improvement it is still not as good as can be. This is a profile with this PR applied. Still there is a lot of iterations of the propagation code (12,000000).

And actually the BVH seems to be taking some time which shouldn't be happening, will investigate - ah .. this is a false problem, it is because this PR is without #45576 applied which cures that. :grin: 

![dynamic_visibility_profile](https://user-images.githubusercontent.com/21999379/106360709-f4f19900-6311-11eb-856a-d358fbe0c9ca.png)

Ideally though we would like to be able to get to a hidden node, and disable propagation of transforms below that node in the scene tree. This is not (afaik) currently supported, presumably because even though child nodes of a hidden node may not need to be visible, they still may need the transform (e.g. physics objects).

I'm currently investigating whether it might be worth us having a single, or set of `disable` modes to prevent propagation of transforms, or perhaps other propagated features.

Any comments welcome. The flags BTW is because it is clear from the node source that there are a number of bools that are intended to be converted to flags but no one has got round to it, so I'm potentially getting the ball rolling on that idea. You could either have access functions for the flags or access the data directly to use bit operations, whichever is preferable. Maybe the latter actually having seen this pattern used in other parts of Godot. Or we could use a bool in this PR, and leave all the conversions to flags in another PR - whichever is preferred.

<!--
Pull requests should always be made for the `master` branch first, as that's
where development happens and the source of all future stable release branches.

Relevant fixes are cherry-picked for stable branches as needed.

Do not create a pull request for stable branches unless the change is already
available in the `master` branch and it cannot be easily cherry-picked.
Alternatively, if the change is only relevant for that branch (e.g. rendering
fixes for the 3.2 branch).
-->