@lawnjelly 2021-02-03
Moved this from draft now as I think it should be a good idea, as it gives a free speed boost and is fairly non-intrusive. That is on the assumption that the visual server doesn't need transforms updated for hidden objects, need to get reduz to confirm this. If he thinks this is the case we can do something similar in master.