@Calinou 2021-02-03
> Ideally though we would like to be able to get to a hidden node, and disable propagation of transforms below that node in the scene tree. This is not (afaik) currently supported, presumably because even though child nodes of a hidden node may not need to be visible, they still may need the transform (e.g. physics objects).

Node2D and Spatial have a `toplevel` property. If `true`, the node won't inherit from its parents' transforms. Can this be used in this optimization?

However, as you state, there's no "parent" counterpart that would make a node's children ignore the parent' transform automatically.