@lawnjelly 2021-02-03
I'll just have a little look to see whether this can be done visual server side, as reduz says he prefers that if possible. :+1: Same kind of idea, just done on the other side of the call.