@lawnjelly 2021-02-03
> Node2D and Spatial have a `toplevel` property (`top_level` in `master`). If `true`, the node won't inherit from its parents' transforms. Can this property be taken into account for this optimization?

I didn't know this property. I'm not absolutely sure if it can help.

> However, as you state, there's no "parent" counterpart that would make a node's children ignore the parent' transform automatically.

The wider problem comes down to this:

* you have a parent, which may have 10,000 children which are all hidden, none of which need their transforms updating.
* You get a propagate transform notification .. it gets sent to 10,000 children (slow)
* 10,000 children either decide they are hidden and do no else (this PR)
* or 10,000 children update their global transform by doing matrix multiplies from the parent, and send this transform to the visual server, which stores it, and then does nothing

Where ideally:
* you have a parent, which has 10,000 children all of which are hidden
* you get a propagate transform notification. You don't send to any children. Done.

In the general case you can't take this latter approach, because you don't know in advance which are hidden, and whether some are nodes that require transform even when hidden (e.g. physics). If you query 10,000 nodes to find this info, you've also lost a load of performance, just by hitting all those nodes unnecessarily.

So ideally you'd want some kind of flags to be able to stop propagating certain things:
* transforms
* process
* physics_process
* etc

I think this is partly what is in this proposal:
https://github.com/godotengine/godot-proposals/issues/1835

Another approach would be to maintain multiple lists on each node, a list of children, and a list of children that need specific updates, and have the children register / unregister. But this becomes quite complex to the point it's not worth it. You could also cache a list of children that need specific updates, and just set these to dirty when things change. That is maybe more feasible.

Having disable flags is probably the most efficient of all but it requires user to take responsibility, whereas some of the others are automatic methods.

There's also the question of where you draw the line with these optimizations, the benefit has to be worth the added code complexity. It is a bit of a trade off. Godot normally errs on the side of simple maintainable code at the cost of some performance.