@lawnjelly 2021-02-03
Ok I've benchmarked both methods now, with 10,000 moving objects as before.

_Current Godot 3.2 dev_
156 fps

_This PR_
395 fps

_Visual Server approach_
207 fps

As I suspected the speed of this PR has increased now that the BVH visibility PR #45576 has been merged.

What is also clear is that the client side approach is not just a bit faster, it is a *lot* faster than bailing out in the visual server. So on balance I am inclined to think this is worth the slight less elegance over visual server approach. What do you guys think?